libs/agentic-common/pyproject.toml
[tool.poetry]
name = "agentic-common"
version = "0.1.0"
description = "Shared utilities for Agentic Bridge services"
authors = ["Agentic Bridge Team"]
packages = [{include = "agentic_common", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
structlog = "^24.1.0"
pyjwt = ">=2.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
